Moved propulsion, except batteries
