+++ openconcept/components/__init__.py
@@ -1,6 +1,2 @@
 from .battery import SimpleBattery
-from .generator import SimpleGenerator
-from .motor import SimpleMotor
-from .propeller import SimplePropeller
-from .splitter import PowerSplit, FlowSplit, FlowCombine
-from .turboshaft import SimpleTurboshaft
+from .manifold import FlowSplit, FlowCombine
