+++ openconcept/examples/B738_VLM_drag.py
@@ -23,7 +23,7 @@ import openconcept.api as oc
 from openconcept.mission import MissionWithReserve, IntegratorGroup
 from openconcept.aerodynamics import OASDragPolar
 from openconcept.examples.aircraft_data.B738 import data as acdata
-from openconcept.components.cfm56 import CFM56
+from openconcept.propulsion import CFM56
 
 class B738AirplaneModel(IntegratorGroup):
     """
