+++ openconcept/propulsion/__init__.py
@@ -1,0 +1,7 @@
+from .cfm56 import CFM56
+from .generator import SimpleGenerator
+from .motor import SimpleMotor
+from .N3 import N3, N3Hybrid
+from .propeller import SimplePropeller, WeightCalc, ThrustCalc, PropCoefficients
+from .splitter import PowerSplit
+from .turboshaft import SimpleTurboshaft
