+++ openconcept/propulsion/empirical_data/cfm56/cfm56smn.npy
