+++ openconcept/propulsion/empirical_data/cfm56/cfm56t4.npy
