+++ openconcept/propulsion/empirical_data/n+3/power_off/T4.npy
