+++ openconcept/propulsion/empirical_data/n+3/power_off/thrust.npy
