+++ openconcept/propulsion/empirical_data/n+3_hybrid/power_off/alt.npy
