+++ openconcept/propulsion/empirical_data/n+3_hybrid/power_on_500kW/wf.npy
