+++ openconcept/propulsion/empirical_data/prop_maps.py
