+++ openconcept/propulsion/empirical_data/propmap_exetended.png
