+++ openconcept/propulsion/generator.py
