+++ openconcept/propulsion/motor.py
