+++ openconcept/propulsion/propeller.py
