+++ openconcept/propulsion/turboshaft.py
